+++ packages/web/src/components/nav/desktop/PlaylistLibrary.tsx
@@ -1,21 +1,13 @@
-import React, { useCallback, useState } from 'react'
+import React, { useCallback } from 'react'
 
-import {
-  IconCaretRight,
-  IconFolder,
-  IconKebabHorizontal,
-  IconFolderOutline
-} from '@audius/stems'
 import cn from 'classnames'
 import { isEmpty } from 'lodash'
 import { useDispatch } from 'react-redux'
-import { NavLink, NavLinkProps } from 'react-router-dom'
 
 import { useModalState } from 'common/hooks/useModalState'
 import { Name } from 'common/models/Analytics'
 import { SmartCollection } from 'common/models/Collection'
 import { ID } from 'common/models/Identifiers'
-import { PlaylistLibraryFolder } from 'common/models/PlaylistLibrary'
 import { SmartCollectionVariant } from 'common/models/SmartCollectionVariant'
 import { FeatureFlags } from 'common/services/remote-config'
 import { AccountCollection } from 'common/store/account/reducer'
@@ -25,12 +17,8 @@ import {
   getPlaylistLibrary
 } from 'common/store/account/selectors'
 import { addTrackToPlaylist } from 'common/store/cache/collections/actions'
-import Draggable from 'components/dragndrop/Draggable'
 import Droppable from 'components/dragndrop/Droppable'
-import IconButton from 'components/icon-button/IconButton'
 import { getPlaylistUpdates } from 'components/notification/store/selectors'
-import Tooltip from 'components/tooltip/Tooltip'
-import UpdateDot from 'components/update-dot/UpdateDot'
 import { useArePlaylistUpdatesEnabled, useFlag } from 'hooks/useRemoteConfig'
 import { SMART_COLLECTION_MAP } from 'pages/smart-collection/smartCollections'
 import { make, useRecord } from 'store/analytics/actions'
@@ -40,303 +28,12 @@ import { getIsDragging } from 'store/dragndrop/selectors'
 import { reorderPlaylistLibrary } from 'store/playlist-library/helpers'
 import { update } from 'store/playlist-library/slice'
 import { useSelector } from 'utils/reducer'
-import { playlistPage, getPathname } from 'utils/route'
+import { getPathname, playlistPage } from 'utils/route'
 
 import navColumnStyles from './NavColumn.module.css'
+import { PlaylistFolderNavItem } from './PlaylistFolderNavItem'
 import styles from './PlaylistLibrary.module.css'
-
-type PlaylistNavLinkProps = NavLinkProps & {
-  droppableKey: ID | SmartCollectionVariant
-  playlistId: ID | SmartCollectionVariant
-  name: string
-  onReorder: (
-    draggingId: ID | SmartCollectionVariant,
-    droppingId: ID | SmartCollectionVariant
-  ) => void
-  link?: string
-}
-
-type PlaylistFolderNavButtonProps = React.ComponentPropsWithoutRef<'button'> & {
-  onReorder: () => void
-}
-
-const PlaylistNavLink = ({
-  droppableKey,
-  playlistId,
-  name,
-  link,
-  onReorder,
-  children,
-  className,
-  ...navLinkProps
-}: PlaylistNavLinkProps) => {
-  const [isDragging, setIsDragging] = useState(false)
-  const onDrag = useCallback(() => {
-    setIsDragging(true)
-  }, [setIsDragging])
-  const onDrop = useCallback(() => {
-    setIsDragging(false)
-  }, [setIsDragging])
-  return (
-    <Droppable
-      key={droppableKey}
-      className={styles.droppable}
-      hoverClassName={styles.droppableHover}
-      onDrop={(id: ID | SmartCollectionVariant) => onReorder(id, playlistId)}
-      acceptedKinds={['library-playlist']}
-    >
-      <Draggable
-        id={playlistId}
-        text={name}
-        link={link}
-        kind='library-playlist'
-        onDrag={onDrag}
-        onDrop={onDrop}
-      >
-        <NavLink
-          {...navLinkProps}
-          draggable={false}
-          className={cn(className, styles.navLink, {
-            [styles.dragging]: isDragging
-          })}
-        >
-          {children}
-        </NavLink>
-      </Draggable>
-    </Droppable>
-  )
-}
-
-const FolderNavLink = ({
-  id,
-  name,
-  onReorder,
-  children,
-  className,
-  ...buttonProps
-}: PlaylistFolderNavButtonProps) => {
-  const [isDragging, setIsDragging] = useState(false)
-  const onDrag = useCallback(() => {
-    setIsDragging(true)
-  }, [setIsDragging])
-  const onDrop = useCallback(() => {
-    setIsDragging(false)
-  }, [setIsDragging])
-
-  return (
-    <Droppable
-      key={id}
-      className={styles.droppable}
-      hoverClassName={styles.droppableHover}
-      onDrop={(id: ID | SmartCollectionVariant) => onReorder()}
-      acceptedKinds={['library-playlist', 'playlist-folder']}
-    >
-      <Draggable
-        id={id}
-        text={name}
-        kind='playlist-folder'
-        onDrag={onDrag}
-        onDrop={onDrop}
-      >
-        <button
-          {...buttonProps}
-          draggable={false}
-          className={cn(className, styles.folderButton, styles.navLink, {
-            [styles.dragging]: isDragging
-          })}
-        >
-          {children}
-        </button>
-      </Draggable>
-    </Droppable>
-  )
-}
-
-const PlaylistFolderNavItem = ({
-  folder,
-  hasUpdate = false,
-  dragging,
-  draggingKind,
-  onClickEdit
-}: {
-  folder: PlaylistLibraryFolder
-  hasUpdate: boolean
-  dragging: boolean
-  draggingKind: string
-  onClickEdit: (folderId: string) => void
-}) => {
-  const { id, name, contents } = folder
-  const isDroppableKind =
-    draggingKind === 'track' ||
-    draggingKind === 'playlist' ||
-    draggingKind === 'playlist-folder'
-  const [isHovering, setIsHovering] = useState(false)
-
-  return (
-    <Droppable
-      key={id}
-      className={navColumnStyles.droppable}
-      hoverClassName={navColumnStyles.droppableHover}
-      onDrop={() => {}}
-      acceptedKinds={['library-playlist']}
-    >
-      <FolderNavLink
-        onMouseEnter={() => {
-          setIsHovering(true)
-        }}
-        onMouseLeave={() => setIsHovering(false)}
-        id={id}
-        name={name}
-        onReorder={() => {}}
-        className={cn(navColumnStyles.link, {
-          [navColumnStyles.droppableLink]: dragging && isDroppableKind,
-          [navColumnStyles.disabledLink]:
-            dragging && !isDroppableKind && draggingKind !== 'library-playlist'
-        })}
-        onClick={() => {}}
-      >
-        <div className={styles.folderButtonContentContainer}>
-          {isEmpty(contents) ? (
-            <IconFolderOutline
-              width={12}
-              height={12}
-              className={styles.iconFolder}
-            />
-          ) : (
-            <IconFolder
-              width={12}
-              height={12}
-              className={cn(styles.iconFolder, {
-                [styles.iconFolderUpdated]: hasUpdate
-              })}
-            />
-          )}
-          <div className={styles.folderNameContainer}>
-            <span>{name}</span>
-          </div>
-          <IconCaretRight height={11} width={11} className={styles.iconCaret} />
-          <IconButton
-            className={cn(styles.iconKebabHorizontal, {
-              [styles.hidden]: !isHovering || dragging
-            })}
-            icon={<IconKebabHorizontal height={11} width={11} />}
-            onClick={e => {
-              e.preventDefault()
-              e.stopPropagation()
-              onClickEdit(id)
-            }}
-          />
-        </div>
-      </FolderNavLink>
-
-      {/* Loop over contents and render playlist list */}
-    </Droppable>
-  )
-}
-
-type PlaylistNavItemProps = {
-  playlist: AccountCollection
-  url: string
-  addTrack: (trackId: ID) => void
-  isOwner: boolean
-  onReorder: (
-    draggingId: ID | SmartCollectionVariant,
-    droppingId: ID | SmartCollectionVariant
-  ) => void
-  hasUpdate?: boolean
-  dragging: boolean
-  draggingKind: string
-  onClickPlaylist: (id: ID, hasUpdate: boolean) => void
-  onClickEdit?: (id: ID) => void
-}
-const PlaylistNavItem = ({
-  playlist,
-  hasUpdate = false,
-  url,
-  addTrack,
-  isOwner,
-  onReorder,
-  dragging,
-  draggingKind,
-  onClickPlaylist,
-  onClickEdit
-}: PlaylistNavItemProps) => {
-  const { id, name } = playlist
-  const [isHovering, setIsHovering] = useState(false)
-
-  return (
-    <Droppable
-      key={id}
-      className={navColumnStyles.droppable}
-      hoverClassName={navColumnStyles.droppableHover}
-      onDrop={addTrack}
-      acceptedKinds={['track']}
-      disabled={!isOwner}
-    >
-      <PlaylistNavLink
-        droppableKey={id}
-        playlistId={id}
-        name={name}
-        link={url}
-        to={url}
-        onReorder={onReorder}
-        isActive={() => url === getPathname()}
-        activeClassName='active'
-        className={cn(navColumnStyles.link, {
-          [navColumnStyles.playlistUpdate]: hasUpdate,
-          [navColumnStyles.droppableLink]:
-            isOwner &&
-            dragging &&
-            (draggingKind === 'track' || draggingKind === 'playlist'),
-          [navColumnStyles.disabledLink]:
-            dragging &&
-            ((draggingKind !== 'track' &&
-              draggingKind !== 'playlist' &&
-              draggingKind !== 'library-playlist') ||
-              !isOwner)
-        })}
-        onClick={() => onClickPlaylist(id, hasUpdate)}
-        onMouseEnter={() => {
-          setIsHovering(true)
-        }}
-        onMouseLeave={() => setIsHovering(false)}
-      >
-        <div className={navColumnStyles.playlistLinkContentContainer}>
-          {!hasUpdate ? null : (
-            <div className={navColumnStyles.updateDotContainer}>
-              <Tooltip
-                className={navColumnStyles.updateDotTooltip}
-                shouldWrapContent={true}
-                shouldDismissOnClick={false}
-                mount={null}
-                mouseEnterDelay={0.1}
-                text='Recently Updated'
-              >
-                <div>
-                  <UpdateDot />
-                </div>
-              </Tooltip>
-            </div>
-          )}
-          <span>{name}</span>
-          {!isOwner || !onClickEdit ? null : (
-            <IconButton
-              className={cn(styles.iconKebabHorizontal, {
-                [styles.hidden]: !isHovering || dragging
-              })}
-              icon={<IconKebabHorizontal height={11} width={11} />}
-              onClick={e => {
-                e.preventDefault()
-                e.stopPropagation()
-                onClickEdit(id)
-              }}
-            />
-          )}
-        </div>
-      </PlaylistNavLink>
-    </Droppable>
-  )
-}
+import { PlaylistNavItem, PlaylistNavLink } from './PlaylistNavItem'
 
 type PlaylistLibraryProps = {
   onClickNavLinkWithAccount: () => void
